AvitoScraping/requirements.txt
requests
beautifulsoup4
psycopg2-binary
python-dotenv
schedule
lxml
fake-useragent
tqdm
retry
